Add integration and unit tests for AMP, data pipelines, and learning rate schedulers
- Implemented tests for AMP scheduler checkpointing in `test_amp_scheduler_checkpoint.py`.
- Added image folder training with AMP in `test_imagefolder_amp_training.py`.
- Created unit tests for AMP functionality in `test_amp.py`.
- Developed tests for file data pipelines, including CSV and NPY datasets in `test_file_data_pipeline.py`.
- Added unit tests for learning rate schedulers in `test_lr_scheduler.py`.
- Included tests for release metadata consistency in `test_release_metadata.py`.

+++ src/mytorch/nn/functional/layers.py
@@ -27,7 +27,7 @@ def linear(input: Tensor, weight: Tensor, bias: Tensor | None = None) -> Tensor:
                 f"linear bias must have shape {(weight.shape[0],)}, "
                 f"got {getattr(bias, 'shape', None)}"
             )
-        result = result + bias
+        result = result + bias.to(dtype=result.dtype)
     return result
 
 
@@ -71,7 +71,7 @@ def bilinear(
     if bias is not None:
         if bias.shape != (weight.shape[0],):
             raise ValueError("bilinear bias has the wrong shape")
-        result = result + bias
+        result = result + bias.to(dtype=result.dtype)
     return result
